Remove unused ledger actions arg at exchange processing history
Also test that for an exchange that generates ledger actions during
history processing they actually get processed.
