+++ rotkehlchen/tests/utils/history.py
@@ -43,9 +43,15 @@ TEST_END_TS = 1559427707
 prices = {
     'USD': {
         'EUR': {
+            1459024920: FVal('0.8982'),
             1467279735: FVal('0.9004'),
+            1485895742: FVal('0.9327'),
+            1500705839: FVal('0.8547'),
+            1519001640: FVal('0.8071'),
             1539713238: FVal('0.8612'),
             1609537953: FVal('0.82411'),
+            1609877514: FVal('0.812'),
+            1611426233: FVal('0.8218'),
         },
     },
     'EUR': {
@@ -89,6 +95,7 @@ prices = {
             1495969504: FVal(1964.685),
             1498694400: FVal(2244.465),
             1500595200: FVal('2295.68'),
+            1500705839: FVal('2424.4'),
             1512561941: FVal(10929.925),
             1512693374: FVal(14415.365),
             1514937600: FVal('12786.04'),
@@ -116,6 +123,7 @@ prices = {
             1446979735: FVal(0.8583),
             1448994442: FVal(0.83195),
             1457279735: FVal(1),
+            1459024920: FVal('9.875'),
             1461021812: FVal('7.875'),
             1463184190: FVal(9.187),
             1463508234: FVal(10.785),
@@ -140,6 +148,7 @@ prices = {
             1539713117: FVal(178.615),
             1539713237: FVal(178.615),
             1539713238: FVal(178.615),
+            1566726126: FVal('167.58'),
             1569924574: FVal('161.59'),
             1607727600: FVal('449.68'),
             1607814000: FVal('469.82'),
